// Client/src/app/models/employeeData.ts
import { City } from './city';
import { State } from './state';

export class EmployeeData {
    id: number;
    noEmployee: number;
    name: string;
    fatherSurname: string;
    motherSurname: string;
    fixNumber: number;
    cellNumber: number;
    street: string;
    colony: string;
    outsideNumber: string;
    insideNumber: string;
    zipCode: number;
    cityId: number;
    stateId: number;
    photoId: number;
}